+++ azurerm/internal/services/storage/shim/containers.go
@@ -1,4 +1,4 @@
-package resourcemanagershim
+package shim
 
 import (
 	"context"
